// src/app/core/services/model/sessionProfile.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AuthorityTemplateNames } from './authorityTemplateNames';

export interface SessionProfile { 
    accessToken?: string;
    sid?: string;
    userId?: string;
    userUUID?: string;
    name?: string;
    authTime?: number;
    authorityTemplateName?: AuthorityTemplateNames;
    isRequiredToChangePassword?: boolean;
    isMoreThanOnePrivilegedSessionOpen?: boolean;
}